Actualización: ArenitoPlugin dentro de arenito.rs
Se movió el plugin de arenito dentro del mismo archivo
